refactor(engine): Move triggers into main API app
